pyproject.toml
[project]
name = "kalbi2"
version = "0.1.0"
description = "Autonomous Multi-Agent Trading System - Prediction Markets & Equities"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Agent Framework
    "crewai[tools]>=0.80.0",

    # LLM
    "anthropic>=0.40.0",
    "litellm>=1.50.0",

    # Trading APIs
    "alpaca-py>=0.30.0",
    "httpx>=0.27.0",

    # Database
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",

    # Cache
    "redis>=5.0.0",

    # Scheduling
    "apscheduler>=3.10.4",

    # Data & Analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "pandas-ta>=0.3.14b1",

    # Configuration
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",

    # News & Data
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.3",
    "requests>=2.32.0",

    # Notifications
    "python-telegram-bot>=21.0",

    # Dashboard
    "streamlit>=1.37.0",

    # Logging
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "types-requests>=2.32.0",
    "types-redis>=4.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"
